+++ scripts/dcurooms.py
@@ -3,6 +3,7 @@
 
 import datetime
 import sys
+import smtplib
 if sys.version_info[0] < 3:
     from cookielib import LWPCookieJar
 else:
@@ -18,13 +19,16 @@ __copyright__ = 'Copyright (c) 2017 theycallmemac'
 __license__ = 'GPL-3.0'
 
 def setup_options():
-    parser = OptionParser(description='Displays and books rooms around the DCU campus via provided timetable/module details', prog='dcurooms', version='%prog ' + __version__, usage='%prog [option]')
+    parser = OptionParser(description='Displays and books rooms around the DCU campus via provided timetable/module details', prog='dcurooms', version='%prog ' + __version__, usage='%prog [option]')  
+    parser.add_option("-l", "--lookup", action="store_true", help="returns information given a specific room, week, day and hour") 
+    parser.add_option("-b", "--book", action="store_true", help="book a room by providing the room,  DD/MM/YYYY, start time, and end time")  
     parser.add_option("-a", "--available", action="store_true", help="returns only the rooms/labs that are free in a building")
-    parser.add_option("-l", "--lookup", action="store_true", help="returns information given a specific room, week, day and hour")
     parser.add_option("-n", "--now", action="store_true", help="show the status of each room/lab as it is at the current time of checking")
-    parser.add_option("-c", "--computing", action="store_true", help="displays the status of the labs in the School of Computing")
-    parser.add_option("-g", "--grattan", action="store_true", help="displays the status of rooms in the Henry Grattan building")
-    return parser
+    # building options which can be passed 
+    parser.add_option("-L", "--computing", action="store_true", help="displays the status of the labs in the School of Computing")
+    parser.add_option("-C", "--grattan", action="store_true", help="displays the status of rooms in the Henry Grattan building")
+
+    return parser 
 
 def check_arguments(week, day):
     if int(week) not in range(1, 53) or int(day) not in range(1,7):
@@ -62,16 +66,85 @@ def build_timetable(room, week, day, hour):
     browser = StatefulBrowser()
     cookie_jar = LWPCookieJar()
     browser.set_cookiejar(cookie_jar)
-    browser.addheaders = [("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36")]
+    browser.user_agent = [("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36")]
     url = "https://www101.dcu.ie/timetables/feed.php?room=" + room + "&week1=" + week + "&hour=" + str(hour) + "&day=" + day + "&template=location"
     browser.open(url)
     return browser, url
+
+def fill_form(args):
+    browser = StatefulBrowser()
+    cookie_jar = LWPCookieJar()
+    browser.set_cookiejar(cookie_jar)
+    room, date, from_time, to_time = args[0], args[1].split("/"), args[2][:2] + ":" + args[2][2:], args[3][:2] + ":" + args[3][2:]
+    day, month, year = date[0], date[1], date[2]
+    if sys.version_info[0] < 3:
+        name = raw_input("Name of society: ")
+        person = raw_input("Your name: ")
+        email = raw_input("Your email: ")
+        number = raw_input("Your number: ")
+    else:
+        name = input("Name of society: ")
+        person = (input("Your name: ")
+        email = input("Your email: ")
+        number = input("Your number: ")
+    browser.open("http://www.dcu.ie/registry/booking.shtml")
+    browser.select_form(nr=2)
+    browser["submitted[name_of_club_society]"] = name
+    browser["submitted[name_of_person_making_booking]"] = person
+    browser["submitted[contact_telephone_number]"] = number
+    browser["submitted[date_room_required][day]"] = day
+    browser["submitted[date_room_required][month]"] = month
+    browser["submitted[date_room_required][year]"] = year
+    browser["submitted[room_capacity]"] = "18"
+    browser["submitted[description_of_event]"] = "Meeting"
+    browser["submitted[hours_requiredfrom_to]"] = from_time + " - " + to_time
+    browser["submitted[building_room_reference]"] = room
+    browser["submitted[email_address]"] = email
+
+    return browser
 
 def check_room(timetable_url):
     html = get(timetable_url)
     soup = BeautifulSoup(html.text, "lxml")
     tr = soup.select('tr')
     return str(tr[12].getText().strip()) + " -> " + str(tr[14].getText().strip())
+
+def make_booking(form):
+    request = form.request
+    response = form.submit_selected()
+    return "Form submitted successfully."
+
+def draft_email(args):
+    if sys.version_info[0] < 3:
+        gmail_user = raw_input("Your gmail: ")
+        gmail_password = raw_input("Your gmail password: ")
+        name = raw_input("Society name: ")
+        person = raw_input("Your name: ")
+    else:
+        gmail_user = input("Your gmail: ")
+        gmail_password = input("Your gmail password: ")
+        name = input("Society name: ")
+        person = input("Your name: ")
+    FROM = gmail_user
+    TO = 'irene.mcevoy@dcu.ie' if type('irene.mcevoy@dcu.ie') is list else ['irene.mcevoy@dcu.ie']
+    SUBJECT = 'Lab Booking'
+    BODY = "Just wondering if you could book " + args[0] + " on the " + args[1] + " from " + args[2][:2] + ":" + args[2][2:] + " to " + args[3][:2] + ":" + args[3][2:] + " for " + name + ".\n\nThank you,\n" + person + "."
+    message = """From: %s\nTo: %s\nSubject: %s\n\n%s
+    """ % (FROM, ", ".join(TO), SUBJECT, BODY)
+
+    return gmail_user, gmail_password, FROM, TO, message
+
+def send_email(gmail_user, gmail_password, FROM, TO, message):
+    try:
+        server = smtplib.SMTP("smtp.gmail.com", 587)
+        server.ehlo()
+        server.starttls()
+        server.login(gmail_user, gmail_password)
+        server.sendmail(FROM, TO, message)
+        server.close()
+        print("Your email has been sent.")
+    except:
+        print("Email failed to send.")
 
 def main():
     parser = setup_options()
@@ -83,7 +156,41 @@ def main():
     c = ['LG25','LG26','LG27','L101','L114','L125','L128']
     g = ['CG01', 'CG02','CG03','CG04','CG05','CG06','CG11','CG12','CG20','CG68','C166']
     details = sys.argv[2:]
-    if len(details) > 5:
+
+    if options.book == True:
+        if len(details) < 4:
+            print("Booking a room requires more arguments. See the help for details.")
+            sys.exit()
+        else:
+            if details[0] in c:
+                user, password, from_who, to_who, message = draft_email(details)
+                if sys.version_info[0] < 3:
+                    confimation = raw_input("\nIs this the correct information? (y/n): ").lower()
+                else:
+                    confirmation = input("\nIs this the correct information? (y/n): ").lower()
+                if confirmation == "y":
+                    send_email(user, password, from_who, to_who, message)
+                else:
+                    print("Draft withdrawn.")
+                sys.exit()
+
+            elif details[0] in g:
+                form = fill_form(details)
+                if sys.version_info[0] < 3:
+                    confimation = raw_input("\nIs this the correct information? (y/n): ").lower()
+                else:
+                    confirmation = input("\nIs this the correct information? (y/n): ").lower()
+                if confirmation == "y":
+                    room_booked = make_booking(form)
+                    print(room_booked)
+                else:
+                    print("Form submission withdrawn.")
+                sys.exit()
+            else:
+                print("That room is not supported by this tool.")
+                sys.exit()
+
+    elif len(details) > 5:
         print("Too many arguments passed.")
         sys.exit()
     elif options.lookup == True: